+++ terraform/node_resource_refresh_test.go
@@ -40,11 +40,13 @@ func TestNodeRefreshableManagedResourceDynamicExpand_scaleOut(t *testing.T) {
 		},
 	}).SyncWrapper()
 
+	cfgAddr := addrs.RootModule.Resource(addrs.ManagedResourceMode, "aws_instance", "foo")
 	n := &NodeRefreshableManagedResource{
 		NodeAbstractResource: &NodeAbstractResource{
-			Addr:   addrs.RootModule.Resource(addrs.ManagedResourceMode, "aws_instance", "foo"),
+			Addr:   cfgAddr,
 			Config: m.Module.ManagedResources["aws_instance.foo"],
 		},
+		Addr: cfgAddr.Absolute(addrs.RootModuleInstance),
 	}
 
 	g, err := n.DynamicExpand(&MockEvalContext{
@@ -120,11 +122,13 @@ func TestNodeRefreshableManagedResourceDynamicExpand_scaleIn(t *testing.T) {
 		},
 	}).SyncWrapper()
 
+	cfgAddr := addrs.RootModule.Resource(addrs.ManagedResourceMode, "aws_instance", "foo")
 	n := &NodeRefreshableManagedResource{
 		NodeAbstractResource: &NodeAbstractResource{
-			Addr:   addrs.RootModule.Resource(addrs.ManagedResourceMode, "aws_instance", "foo"),
+			Addr:   cfgAddr,
 			Config: m.Module.ManagedResources["aws_instance.foo"],
 		},
+		Addr: cfgAddr.Absolute(addrs.RootModuleInstance),
 	}
 
 	g, err := n.DynamicExpand(&MockEvalContext{
@@ -160,14 +164,10 @@ func TestNodeRefreshableManagedResourceEvalTree_scaleOut(t *testing.T) {
 	m := testModule(t, "refresh-resource-scale-inout")
 
 	n := &NodeRefreshableManagedResourceInstance{
-		NodeAbstractResourceInstance: &NodeAbstractResourceInstance{
-			NodeAbstractResource: NodeAbstractResource{
-				Addr:   addrs.RootModule.Resource(addrs.ManagedResourceMode, "aws_instance", "foo"),
-				Config: m.Module.ManagedResources["aws_instance.foo"],
-			},
-			InstanceKey: addrs.IntKey(2),
-		},
+		NodeAbstractResourceInstance: NewNodeAbstractResourceInstance(addrs.RootModuleInstance.Resource(addrs.ManagedResourceMode, "aws_instance", "foo").Instance(addrs.IntKey(2))),
 	}
+
+	n.AttachResourceConfig(m.Module.ManagedResources["aws_instance.foo"])
 
 	actual := n.EvalTree()
 	expected := n.evalTreeManagedResourceNoState()